refractor: FE Overviewpage -> move blendprogress to timeUtil
